Added a very important optimization for calling functions.
Now, if your function (user-defined or built-in) has i <= 4
parameters, it is optimized as follows.

1. When a scheme object is made from it, instead of
creating a general Lambda object (which means that at calling
time its parameters will be made into a cons, which will
have to be deconstructed inside the fucntion, effectively
doing the same work twice for nothing!), a special
Lambdai object is creatd that holds the actual implemntation
of the function (as a function of i ocaml parameters).

2. On the calling isde, special functions applyi have been
created. When a variable, supposedly containing a function
is called with i parameters, the special applyi functions is
generated which will check that the object is in fact a
Lambdai object and dispatch the parameters directly, no need
to make them into a cons just to deconstruct the cons on
the other side!

3. The only moment in which this doesn't work is when
we cannot guess the right number of parameters before hand
(e.g. var args). Also we can optimize a little more the
calling of user-defined functions which are not modified
(with set!) by not even creating the Lambda/Lambdai object
and just calling them directly in Ocaml (the arity can
be checked statically in this case).