Started work on macros.
Now the Ast module will handle macro expansion and scope
analysis, but we don't do anything for undefined variables,
as we will treat that later (to allow for recursive definitions
on toplevel).
A macro will be a transformation datum -> datum, while a special
form is a transformation datum -> pexp. Have a preliminary
macro matcher in macro.ml, but it doesn't handle '...' correctly.